requirements.txt
streamlit
matplotlib
pillow
numpy